requirements.txt
asgiref==3.2.7
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
cssselect==1.1.0
cssutils==1.0.2
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-embed-video==1.3.2
django-imagekit==4.0.2
django-js-asset==1.2.2
django-render-block==0.6
django-templated-email==2.3.0
html2text==2020.1.16
idna==2.9
lxml==4.5.1
pilkit==2.0
Pillow==7.1.2
premailer==3.7.0
psycopg2==2.7.5
psycopg2-binary==2.8.5
pytz==2019.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.8
